Analitica_Python/requirements.txt
fastapi
uvicorn[standard]
pandas
requests
prophet
sqlalchemy
psycopg2-binary
python-dotenv